requirements/requirements_base.txt
Django>=2.1.7,<3
djangorestframework>=3.9.2,<4
django-haystack>=2.8.1,<3
elasticsearch>=5.5.3,<6
django-haystack-elasticsearch>=0.1.0,<1
